letterWheel.json
JSON used for the HQ Words letter wheel. Originally we were going to implement a practice spin mode where a user would be able to test a spin before having a letter be credited. Also, we shortly implemented a mode for multi spins which would allow a user to keep the letter they landed on or spin again for a chance at getting a better letter