countdown
In early 2018, HQ began to experiment with a real-time rendered animation of their famous 2:18 countdown. Behind the animation, HQ would show a graphic on stream that said "Countdown in place, the game will begin in xx seconds, please head to your assigned area". The animation has been played in 17 internal games with 4 variants of the animation being used